add function to save and edit project
